# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Clean Architecture Go Gin API",
        "title": "Go Gin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Afandi",
            "email": "mohamadafandi71@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/api/v1",
    "paths": {
...
        "/recurring": {
            "get": {
                "description": "Get all recurring transactions for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring"
                ],
                "summary": "Get recurring transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessRecurringResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new recurring transaction template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring"
                ],
                "summary": "Create recurring transaction",
                "parameters": [
                    {
                        "description": "Recurring Transaction payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recurring_transaction.RecurringTransaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/recurring/process": {
            "post": {
                "description": "Trigger the manual processing of due recurring transactions into transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring"
                ],
                "summary": "Process due recurring transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/recurring/{id}": {
            "delete": {
                "description": "Delete an existing recurring transaction template",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring"
                ],
                "summary": "Delete recurring transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recurring Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/reports/spending": {
            "get": {
                "description": "Get a breakdown of spending by category for a specific month and year",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get category spending report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Month (1-12)",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "recurring_transaction.Frequency": {
            "type": "string",
            "enum": [
                "daily",
                "weekly",
                "monthly"
            ],
            "x-enum-varnames": [
                "Daily",
                "Weekly",
                "Monthly"
            ]
        },
        "recurring_transaction.RecurringTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "frequency": {
                    "$ref": "#/definitions/recurring_transaction.Frequency"
                },
                "id": {
                    "type": "integer"
                },
                "last_processed": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "description": "\"income\" or \"expense\"",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorSwaggerResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "response.SuccessRecurringResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/recurring_transaction.RecurringTransaction"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}